[Gearpump 311] refactor state management
