+++ streaming/src/main/java/org/apache/gearpump/streaming/refactor/coder/ElementByteSizeObserver.java
@@ -1,0 +1,66 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.apache.gearpump.streaming.refactor.coder;
+
+import java.util.Observable;
+import java.util.Observer;
+
+public abstract class ElementByteSizeObserver implements Observer {
+    private boolean isLazy = false;
+    private long totalSize = 0;
+    private double scalingFactor = 1.0;
+
+    public ElementByteSizeObserver() {
+    }
+
+    protected abstract void reportElementSize(long elementByteSize);
+
+    public void setLazy() {
+        isLazy = true;
+    }
+
+    public boolean getIsLazy() {
+        return isLazy;
+    }
+
+    public void update(Object obj) {
+        update(null, obj);
+    }
+
+    public void setScalingFactor(double scalingFactor) {
+        this.scalingFactor = scalingFactor;
+    }
+
+    @Override
+    public void update(Observable obs, Object obj) {
+        if (obj instanceof Long) {
+            totalSize += scalingFactor * (Long) obj;
+        } else if (obj instanceof Integer) {
+            totalSize += scalingFactor * (Integer) obj;
+        } else {
+            throw new AssertionError("unexpected parameter object");
+        }
+    }
+
+    public void advance() {
+        reportElementSize(totalSize);
+        totalSize = 0;
+        isLazy = false;
+    }
+}
