+++ streaming/src/main/java/org/apache/gearpump/streaming/refactor/coder/StructuralByteArray.java
@@ -22,9 +22,6 @@ import com.google.common.io.BaseEncoding;
 
 import java.util.Arrays;
 
-/**
- * a structural byte array inspired by Apache Beam
- */
 public class StructuralByteArray {
     byte[] value;
