+++ streaming/src/main/java/org/apache/gearpump/streaming/refactor/coder/VarLongCoder.java
@@ -22,11 +22,6 @@ import java.io.*;
 import java.util.Collections;
 import java.util.List;
 
-/**
- * A {@link Coder} that encodes {@link Long Longs} using between 1 and 10 bytes. Negative
- * numbers always take 10 bytes, so {@link BigEndianLongCoder} may be preferable for
- * longs that are known to often be large or negative.
- */
 public class VarLongCoder extends StructuredCoder<Long> {
     public static VarLongCoder of() {
         return INSTANCE;
@@ -73,21 +68,11 @@ public class VarLongCoder extends StructuredCoder<Long> {
     @Override
     public void verifyDeterministic() {}
 
-    /**
-     * {@inheritDoc}
-     *
-     * @return {@code true}. {@link VarLongCoder} is injective.
-     */
     @Override
     public boolean consistentWithEquals() {
         return true;
     }
 
-    /**
-     * {@inheritDoc}
-     *
-     * @return {@code true}. {@link #getEncodedElementByteSize} is cheap.
-     */
     @Override
     public boolean isRegisterByteSizeObserverCheap(Long value) {
         return true;
